extract machine
